Rename the after emit logger
